move proxy graph implementation into a separate file
